requirements.txt
# django packages
Django==3.2.13
djangorestframework==3.12.2
django-cors-headers==3.1.0
dj-rest-auth==2.2.4
djangorestframework-simplejwt==5.2.0
django-storages==1.11.1
django-extensions==3.1.3
django_celery_beat==2.2.0
django_celery_results==2.2.0
django-redis==5.0.0
django-environ==0.4.5
django-filter==2.4.0
drf-nested-routers==0.93.3
django-reversion==4.0.0
django-reversion-compare==0.14.1
django-notifications-hq==1.6.0
django-import-export==2.6.1
django-ipware==4.0.0
django-taggit==2.0.0
django-admin-multiple-choice-list-filter==0.1.1
django-admin-rangefilter==0.8.3
django-anymail==8.5
django-rest-passwordreset==1.2.1
django-ses==2.4.0
django-autocomplete-light==3.9.4

# channels
channels==3.0.4
channels-redis==3.3.1

# aws
boto3==1.17.100

# web
gunicorn==20.1.0

# error
sentry-sdk==1.1.0

# celery
celery==5.1.0
redis==3.5.3

# libs
requests==2.25.1
psycopg2-binary>=2.8,<2.9
openpyxl==2.6.3
imap-tools==0.41.0
Pillow==9.0.1
tabulate==0.8.9
tqdm==4.61.0
six==1.16.0

# google
google-api-python-client==1.11.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.4
